# app02-api-docs/swagger.yaml
openapi: "3.0.0"
info:
  title: learn express & swagger
  description: ---
  version: 1.1.0
  contact:
    email: introbond.dev@gmail.com
    url: "https://introbond-awards.web.app"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for localhost - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for localhost -regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for localhost - secure
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /core:
    get:
      tags: 
        - String
      summary: returns a greet message from core
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
        400:
          description: Bad request
        500:
          description: internal server error

  /object:
    get:
      tags: 
        - Object
      summary: returns a unique course
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad request
        500:
          description: internal server error

  /courses:
    get:
      tags: 
        - Array
      summary: returns all courses
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad request
        500:
          description: internal server error
  /mycourse/{courseId}:
    get:
      tags: 
        - String
      summary: returns course based on request id
      parameters:
        - name: courseId
          in: path
          requried: true
          default: 33
          schema:
            type: string
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
        400:
          description: Bad request
        500:
          description: internal server error

  /addCourse:
    post:
      tags: 
        - String
      summary: add a new corse to existing courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: boolean
        400:
          description: Bad request
        500:
          description: internal server error

  /coursequery:
    get:
      tags: 
        - String
      summary: learn about query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, london, jaipur]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device  :
                    type: string
        400:
          description: Bad request
        500:
          description: internal server error

  /courseupload:
    post:
      tags: 
        - String
      summary: learn about uploading image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: All good success
        400:
          description: Bad request
        500:
          description: internal server error